#49 Remove old process_iteration loop
This is code cleanup done in prep for moving the writing of output into tool implementations.  'process_iteration2' has been used in place of 'process_iteration', but both loops remained in the code.  Now that we have to refactor to remove writing of the draft document, it makes sense to remove 'process_iteration'.  (Otherwise, both would need to be refactored).